requirements.txt
flywheel-sdk==10.6.1
pydicom
click
